Change to adjust for "\auto" commands in YEP Msg1
"YEP Extended Message Pack 1" has spcial commands that reset the width/height of the message bot completely. One of those commands is "\auto" - it recalculates width and height based on the text inputted and, if place in the "center", makes the box much slammer.
This is an issue with the set width/height of the plugin.

Solution: Added a detector for "special" cases where default height/width generation should be used.

This is a bit hacky, though. Should address better when I beef up for "talking faces"

Bug report is here: https://forums.rpgmakerweb.com/index.php?threads/mirror-portraits-in-coversation-plugin-ww-converse.97105/page-2#post-921525